sdk/js: Move address functions to cosmos package
`canonicalAddress` and `humanAddress` are useful functions for all
cosmos chains so move it into a new cosmos package.  This also means we
need to stop hardcoding "terra" in `humanAddress` and take the human
readable part as a parameter instead.
