+++ sdk/js/src/cosmos/index.ts
@@ -1,0 +1,1 @@
+export * from "./address";
