+++ sdk/js/src/index.ts
@@ -1,3 +1,4 @@
+export * from "./cosmos";
 export * from "./ethers-contracts";
 export * from "./solana";
 export * from "./terra";
@@ -6,6 +7,7 @@ export * from "./utils";
 export * from "./bridge";
 export * from "./token_bridge";
 
+export * as cosmos from "./cosmos";
 export * as ethers_contracts from "./ethers-contracts";
 export * as solana from "./solana";
 export * as terra from "./terra";
