+++ sdk/js/src/token_bridge/getOriginalAsset.ts
@@ -7,7 +7,8 @@ import { decodeLocalState } from "../algorand";
 import { buildTokenId } from "../cosmwasm/address";
 import { TokenImplementation__factory } from "../ethers-contracts";
 import { importTokenWasm } from "../solana/wasm";
-import { buildNativeId, canonicalAddress, isNativeDenom } from "../terra";
+import { buildNativeId, isNativeDenom } from "../terra";
+import { canonicalAddress } from "../cosmos";
 import {
   ChainId,
   ChainName,
